requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0 
httpx==0.27.0
gspread==5.12.4
oauth2client==4.1.3
pydantic>=2.4.1,<2.6
google-cloud-translate

openai==1.57.0
# gigachain_community==0.0.19
